requirements.txt
pyautogui>=0.9.54
Pillow>=10.0.0
pystray>=0.19.5
pynput>=1.7.6